Add big brother edit password layout
